Refactor: Organize core modules into 'core' package
- Moved ai_config_manager.py, ai_gemini_handler.py, ai_tool_manager.py,
  and ai_type_definitions.py into a new 'core/' subdirectory.
- Added __init__.py to 'core/' and 'tools/' to define them as Python
  packages.
- Updated import statements in google_ai_integration.py and within
  the core modules to reflect the new structure (using relative
  imports where appropriate within 'core').
- Adjusted path logic in ai_tool_manager.py to correctly locate
  the 'tools/' directory.
- Ensured application remains directly runnable and tested successfully
  with 'info' command and a multi-tool query involving file and
  weather tools.
- Includes minor formatting updates and logging refinements in
  ai_config_manager.py and google_ai_integration.py.

This refactoring improves project organization and modularity.
